Supporting non-full screen sizes and a bug fix
1. Supported non-full screen sizes - Added new CSS, added code to each
of the states to check if the app bar should be enabled, added code to
each of the states to reposition the UI elements if we were launched in
non-full screen state.

2. There were a number of bugs with clicking on the headers in the
search experience.  First was that the WHEN header was never getting
click enabled.  Second was that you could click on any state at any
points.  Third was that when the searchstate got reset, the headers were
not (WRT click handlers).